+++ example/src/data.ts
@@ -46,4 +46,22 @@ const testData: TestData[] = _.range(3000).map(i => ({
   zipCode: randZipCode()
 }));
 
-export { testData, type TestData };
+const generateTestData = (rows: number) => {
+  return _.range(rows).map(i => ({
+    id: i + 1,
+    firstName: randFirstName(),
+    lastName: randLastName(),
+    email: randEmail(),
+    avatar: randImg({ width: 134, height: 134 }).replace("random=", "?random="),
+    country: randCountryCode().toLowerCase(),
+    words: randCatchPhrase(),
+    sentence: randSentence(),
+    lorem: randParagraph(),
+    address: randStreetAddress(),
+    city: randCity(),
+    state: randStateAbbr(),
+    zipCode: randZipCode()
+  }));
+};
+
+export { generateTestData, testData, type TestData };