+++ index.d.ts
@@ -335,6 +335,21 @@ export type TableProps<T> = {
     index: number;
     event?: React.MouseEvent<Element, MouseEvent>;
   }) => void;
+  /**
+   * Function to get more data when you reach the bottom of the table. If this returns
+   * true, then when it reaches the bottom this function will run. If it returns false, then
+   * we have exhausted all remaining data.
+   */
+  onLoadRows?: () => Promise<boolean>;
+  /**
+   * Number of rows from the bottom to trigger loading more rows. This is ignored if
+   * onLoadRows is not specified (default: 1).
+   */
+  asyncOverscan?: number;
+  /**
+   * optional component that gets rendered at the bottom of the table
+   */
+  endComponent?: (data: { isLoading: boolean; hasMoreData: boolean }) => ReactNode;
 };
 
 /**